P2P: Resolve frequency conflicts between STA and P2P (BRCM)
Upon association check whether there's a a frequency
conflict given that no support for multi channel. In case
there's a freq conflict with an existing P2P_GO or P2P_CLI
decide how to resolve it (which one to disable or disconnect)
according to the priority configuration.

Effectively on Android JB, STA is always prioritized over P2P
so in case of a conflict P2P will be disconnected so that STA
can associate on a different channel.

[Igal - fix compilation issue with enum p2p_group_removal_reason]

Signed-off-by: Eyal Shapira <eyal@wizery.com>

Conflicts:

	wpa_supplicant/wpa_supplicant_i.h

Change-Id: I47f0d4d4966e572f1bf8791531498199ffc77613
Signed-off-by: Igal Chernobelsky <igalc@ti.com>